Support detecting modal dismissals
